+++ src/main/java/com/example/psbogroup3/web/model/request/UpdateJobRequest.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class UpdateJobRequest {
     private String institution;
     private String position;
-    private double salary;
+    private long salary;
     private String joinDate;
 
 }
